내서재 UI 구현 및 링크복사 기능 구현

+++ src/components/BookDetails/Comments/CommentList.jsx
@@ -19,7 +19,10 @@ function CommentList(props) {
             <button
               className={css.moreArea}
               onClick={() => {
-                onRemove(id);
+                if (window.confirm('정말 삭제할까요?')) {
+                  alert('성공적으로 삭제되었습니다!');
+                  onRemove(id);
+                }
               }}
             >
               삭제하기
@@ -42,8 +45,17 @@ function CommentList(props) {
         <div className={css.reviewLike}>
           <p>이 리뷰가 마음에 드시나요?</p>
           <button type="button" className={css.likeBtn}>
-            <img src="img/heart.svg" />
-            <span>1</span>
+            <svg
+              xmlns="http://www.w3.org/2000/svg"
+              width="11"
+              height="11"
+              fill="currentColor"
+              class="bi bi-heart"
+              viewBox="0 0 16 16"
+            >
+              <path d="m8 2.748-.717-.737C5.6.281 2.514.878 1.4 3.053c-.523 1.023-.641 2.5.314 4.385.92 1.815 2.834 3.989 6.286 6.357 3.452-2.368 5.365-4.542 6.286-6.357.955-1.886.838-3.362.314-4.385C13.486.878 10.4.28 8.717 2.01L8 2.748zM8 15C-7.333 4.868 3.279-3.04 7.824 1.143c.06.055.119.112.176.171a3.12 3.12 0 0 1 .176-.17C12.72-3.042 23.333 4.867 8 15z" />
+            </svg>
+            <span>0</span>
           </button>
         </div>
       </div>
